requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
clip==0.2.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
diffusers==0.32.1
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.3
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.2
pillow==10.2.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.1
zipp==3.21.0
gradio==4.19.2
pandas==2.2.0